// app/types/emuns.ts
export enum E_TYPE_INPUT {
  BORDER = "border",
}
  
export enum E_TYPE_BUTTON {
  PRIMARY = "primary",
  OUTLINE = "outline",
}

export enum E_TYPE_PAYMENT {
  WALLET = "WALLET",
  CASH = "CASH",
}

export enum E_TYPE_ORDER_STATUS {
  PENDING = "PENDING",
  CANCEL = "CANCEL",
  DELIVERY = "DELIVERING",
  ACCEPTED = "ACCEPTED",
  REJECTED = "REJECTED",
  DONE = "DONE",
  IGNORE = "IGNORE",
}

export enum E_ROLE_ORDER {
  BUY = "BUY",
  SELL = "SELL",
}

export enum ESystemRoles {
  ADMIN = "ADMIN",
  USER = "USER",
}

export enum E_STATUS_WALLET {
  NORMAL = "NORMAL",
  DEPOSIT_PENDING = "DEPOSIT_PENDING",
  WITHDRAW_REQUEST  = "WITHDRAW_REQUEST",
  WITHDRAW_CONFIRM = "WITHDRAW_CONFIRM",
}

export enum E_TYPE_NOTIFICATION {
  USER_CREATED = "USER_CREATED",
  ORDER_CREATED = "ORDER_CREATED",
  ORDER_CANCEL = "ORDER_CANCEL",
  ORDER_ACCEPTED = "ORDER_ACCEPTED",
  ORDER_DELIVERY = "ORDER_DELIVERY",
  ORDER_REJECTED = "ORDER_REJECTED",
  ORDER_DONE = "ORDER_DONE",
  WALLET_REQUIRE_DEPOSIT = "WALLET_REQUIRE_DEPOSIT",
  WALLET_ACCEPTED_DEPOSIT = "WALLET_ACCEPTED_DEPOSIT",
  WALLET_REQUIRE_WITHDRAW = "WALLET_REQUIRE_WITHDRAW",
  WALLET_ACCEPTED_WITHDRAW = "WALLET_ACCEPTED_WITHDRAW",
  WALLET_REFUND = "WALLET_REFUND"
}